+++ jest.config.cjs
@@ -33,6 +33,9 @@ module.exports = {
         '<rootDir>/src/tests/dummy-connector/metadata-extraction.test.ts',
         '<rootDir>/src/http/axios-client-internal.test.ts',
         '<rootDir>/src/tests/event-data-size-limit/size-limit-1.test.ts',
+        '<rootDir>/src/tests/partial-load-failure/partial-load-failure.test.ts',
+        '<rootDir>/src/tests/state-atomicity/post-state-failure.test.ts',
+        '<rootDir>/src/tests/attachment-stream-leak/stream-leak.test.ts',
       ],
     },
   ],
