+++ src/multithreading/process-task.test.ts
@@ -1,11 +1,16 @@
 import { EventType } from '../types/extraction';
 import { WorkerMessageSubject } from '../types/workers';
 
-/* eslint-disable @typescript-eslint/no-require-imports */
-
-// ---------------------------------------------------------------------------
-// Module mocks – all must be declared before any imports that trigger them
-// ---------------------------------------------------------------------------
+// These tests cover logic that is NOT exercised by the end-to-end integration
+// tests under src/tests/timeout-handling/:
+//   - translation of legacy wire event types into the new enum (mutates event in place)
+//   - the hasWorkerEmitted guard that prevents onTimeout from firing after a
+//     successful emit (integration tests only exercise the positive case)
+//   - the error branch that posts WorkerMessageFailed and exits(1)
+//
+// Tests for the happy path, timeout-signal-reaches-worker behavior, and main-thread
+// early return were removed — they either duplicate what the integration suite
+// already exercises or assert mocked behavior with little signal.
 
 const mockParentPortPostMessage = jest.fn();
 const mockParentPortOn = jest.fn();
@@ -58,18 +63,11 @@ jest.mock('./worker-adapter/worker-adapter', () => ({
   })),
 }));
 
-// ---------------------------------------------------------------------------
-// Imports (after mocks)
-// ---------------------------------------------------------------------------
 import { processTask } from './process-task';
 import { translateIncomingEventType } from '../common/event-type-translation';
 import { createAdapterState } from '../state/state';
 import { WorkerAdapter } from './worker-adapter/worker-adapter';
 import { createMockEvent } from '../common/test-utils';
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
 
 function setWorkerData(data: Record<string, unknown>) {
   (global as Record<string, unknown>).__workerData__ = data;
@@ -81,12 +79,8 @@ function makeEvent(eventType = EventType.StartExtractingData) {
   });
 }
 
-// Flush the microtask queue enough to let the async IIFE inside processTask run
-const flush = () => new Promise((r) => setTimeout(r, 0));
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
+// Flush the microtask queue enough to let the async IIFE inside processTask run.
+const flush = async () => new Promise((r) => setTimeout(r, 0));
 
 describe('processTask', () => {
   let processExitSpy: jest.SpyInstance;
@@ -106,19 +100,6 @@ describe('processTask', () => {
     processExitSpy.mockRestore();
   });
 
-  // -------------------------------------------------------------------------
-  it('should return early when running on the main thread', () => {
-    mockIsMainThread = true;
-    const task = jest.fn();
-    const onTimeout = jest.fn();
-
-    processTask({ task, onTimeout });
-
-    expect(task).not.toHaveBeenCalled();
-    expect(onTimeout).not.toHaveBeenCalled();
-  });
-
-  // -------------------------------------------------------------------------
   it('should translate incoming event type before passing to task', async () => {
     const event = makeEvent(EventType.StartExtractingData);
     setWorkerData({ event, initialState: {}, options: {} });
@@ -141,67 +122,13 @@ describe('processTask', () => {
     expect(event.payload.event_type).toBe(EventType.StartExtractingMetadata);
   });
 
-  // -------------------------------------------------------------------------
-  it('should call task with the adapter on the happy path and exit(0)', async () => {
-    const event = makeEvent();
-    setWorkerData({ event, initialState: {}, options: {} });
-
-    const mockAdapter = { isTimeout: false, hasWorkerEmitted: false };
-    (WorkerAdapter as jest.Mock).mockImplementation(() => mockAdapter);
-
-    const task = jest.fn().mockResolvedValue(undefined);
-    const onTimeout = jest.fn().mockResolvedValue(undefined);
-
-    processTask({ task, onTimeout });
-    await flush();
-
-    expect(task).toHaveBeenCalledWith({ adapter: mockAdapter });
-    expect(onTimeout).not.toHaveBeenCalled();
-    expect(processExitSpy).toHaveBeenCalledWith(0);
-  });
-
-  // -------------------------------------------------------------------------
-  it('should call onTimeout after task finishes when a WorkerMessageExit arrives mid-task', async () => {
-    const event = makeEvent();
-    setWorkerData({ event, initialState: {}, options: {} });
-
-    const mockAdapter = { isTimeout: false, hasWorkerEmitted: false };
-    (WorkerAdapter as jest.Mock).mockImplementation(() => mockAdapter);
-
-    // Capture the handler that processTask registers via parentPort.on so we
-    // can fire it ourselves at the right moment — after registration but
-    // before the task resolves.
-    let registeredHandler: ((msg: { subject: string }) => void) | null = null;
-    mockParentPortOn.mockImplementation(
-      (_event: string, cb: (msg: { subject: string }) => void) => {
-        registeredHandler = cb;
-      }
-    );
-
-    const task = jest.fn().mockImplementation(async () => {
-      // At this point the handler is already registered.  Simulate the main
-      // thread sending WorkerMessageExit while the task is running.
-      expect(registeredHandler).not.toBeNull();
-      registeredHandler!({ subject: WorkerMessageSubject.WorkerMessageExit });
-      // isTimeout is now true (set by the production handler), hasWorkerEmitted
-      // is still false — onTimeout should be called after this returns.
-    });
-    const onTimeout = jest.fn().mockResolvedValue(undefined);
-
-    processTask({ task, onTimeout });
-    await flush();
-
-    expect(onTimeout).toHaveBeenCalledWith({ adapter: mockAdapter });
-    expect(processExitSpy).toHaveBeenCalledWith(0);
-  });
-
-  // -------------------------------------------------------------------------
   it('should NOT call onTimeout when the worker already emitted before timeout check', async () => {
     const event = makeEvent();
     setWorkerData({ event, initialState: {}, options: {} });
 
     // Both flags true: a timeout arrived but the worker had already emitted —
-    // onTimeout must be skipped.
+    // onTimeout must be skipped. This is the guard the integration suite cannot
+    // target cleanly because it requires a precise race between emit and timeout.
     const mockAdapter = { isTimeout: true, hasWorkerEmitted: true };
     (WorkerAdapter as jest.Mock).mockImplementation(() => mockAdapter);
 
@@ -215,7 +142,6 @@ describe('processTask', () => {
     expect(processExitSpy).toHaveBeenCalledWith(0);
   });
 
-  // -------------------------------------------------------------------------
   it('should post WorkerMessageFailed with the error message and exit(1) when task throws', async () => {
     const event = makeEvent();
     setWorkerData({ event, initialState: {}, options: {} });
@@ -239,51 +165,6 @@ describe('processTask', () => {
       })
     );
     expect(processExitSpy).toHaveBeenCalledWith(1);
-  });
-
-  // -------------------------------------------------------------------------
-  it('should not call onTimeout when task throws', async () => {
-    const event = makeEvent();
-    setWorkerData({ event, initialState: {}, options: {} });
-
-    const mockAdapter = { isTimeout: false, hasWorkerEmitted: false };
-    (WorkerAdapter as jest.Mock).mockImplementation(() => mockAdapter);
-
-    const task = jest.fn().mockRejectedValue(new Error('fail'));
-    const onTimeout = jest.fn().mockResolvedValue(undefined);
-
-    processTask({ task, onTimeout });
-    await flush();
-
     expect(onTimeout).not.toHaveBeenCalled();
-  });
-
-  // -------------------------------------------------------------------------
-  it('should ignore non-exit worker messages — WorkerMessageLog does not set isTimeout', async () => {
-    const event = makeEvent();
-    setWorkerData({ event, initialState: {}, options: {} });
-
-    const mockAdapter = { isTimeout: false, hasWorkerEmitted: false };
-    (WorkerAdapter as jest.Mock).mockImplementation(() => mockAdapter);
-
-    let registeredHandler: ((msg: { subject: string }) => void) | null = null;
-    mockParentPortOn.mockImplementation(
-      (_event: string, cb: (msg: { subject: string }) => void) => {
-        registeredHandler = cb;
-      }
-    );
-
-    const task = jest.fn().mockImplementation(async () => {
-      // Fire a WorkerMessageLog — should be a no-op for isTimeout
-      registeredHandler!({ subject: WorkerMessageSubject.WorkerMessageLog });
-    });
-    const onTimeout = jest.fn().mockResolvedValue(undefined);
-
-    processTask({ task, onTimeout });
-    await flush();
-
-    expect(mockAdapter.isTimeout).toBe(false);
-    expect(onTimeout).not.toHaveBeenCalled();
-    expect(processExitSpy).toHaveBeenCalledWith(0);
   });
 });
